Rename routes temporarily
