+++ src/routes/Routes.mobile.tsx
@@ -4,14 +4,15 @@ import { withRouter, BrowserRouter, Route, Switch } from "react-router-dom";
 import RootContainer from '@src/containers/RootContainer/RootContainer'
 import DefaultLayout from '@src/layouts/DefaultLayout/DefaultLayout';
 import ErrorLayout from '@src/layouts/ErrorLayout/ErrorLayout';
+import DefaultRoute from './DefaultRoute/DefaultRoute';
 
-export default () => {
+const Routes = () => {
   return (
     <BrowserRouter>
       <RootContainer>
         <Switch>
           <Route path='/error' component={ErrorLayout}/>
-          <Route path='/' component={DefaultLayout}/>
+          <Route path='/' component={DefaultRoute}/>
         </Switch>
     </RootContainer>
    </BrowserRouter>
@@ -19,15 +20,6 @@ export default () => {
 }
 
 /**
- * Router-router navigation test logic
- * DO NOT DELETE
+ * Do not connect redux, otherwise react-hot-loader won't work.
  */
-// const _A = (props) => {
-//   return <div onClick={() => {console.log(1); props.history.push('B')}}>AA</div>
-// }
-// const A = withRouter(_A)
-
-// const _B = (props) => {
-//   return <div onClick={() => {console.log(2); props.history.push('A')}}>BB</div>
-// }
-// const B = withRouter(_B);
+export default Routes;
